requirements.txt
Django==1.7.1
django-colorfield==0.1.3
django-randomslugfield==0.3.0
django-tastypie==0.12.0
jsonfield==1.0.0
jsonschema==2.4.0
python-bugzilla
requests
